requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.1.1
anyio==4.0.0
asgiref==3.7.2
astroid==3.0.1
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.3.2
cron-descriptor==1.4.0
dill==0.3.7
Django==4.2.6
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.3
django-rest-framework==0.1.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
flake8==6.1.0
frozenlist==1.4.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inflection==0.5.1
isort==5.12.0
kombu==5.3.2
mccabe==0.7.0
multidict==6.0.4
packaging==23.2
Pillow==10.0.1
platformdirs==3.11.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pylint==3.0.1
pyTelegramBotAPI==4.14.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==6.7.0
telebot==0.0.5
telepot==12.7
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
